Add files via upload
Dataset consists of 384977 rows and 22 columns. There are 3 float value columns, 10 integer columns and 9 object column.
Here we are dealing with monthly rental prices of different housing types like apartment, condo etc. Also, price of the houses always depends on the amenities which are being provided for the rental by the owner and its locality. So, we understand that the prices are affected by different factors. 
Model Improvization- Since many assumptions are violated and linear model is not working properly so used non-linear model.
We used different algorithms to find the best model for the business problem. The proposed Algorithms are
Decision Tree
 Random Forest
 XGBoost
Adaboost
Gradient boost
After checking all the Assumptions, we can clearly say that every assumptions are violated, so we cannot be able to perform Linear Regression model for this data. Hence, we tried to apply Non-linear models like decision tree, Random Forest, Adaboost, Gradient Boosting and also XGBoost regressor.  Compared to all model , XGBBoost shows 99.89% as  R-square value which is more than all.